Save on demand, make rogue mode optional

+++ src/gameUtils/fadeIntoGameOver.js
@@ -2,7 +2,7 @@ import * as db from '../db'
 import constants from '../constants'
 import { CTDLGAME } from './CTDLGAME'
 import { initSoundtrack, stopMusic, changeVolume } from '../soundtrack'
-import { newGameButton } from '../events'
+import { loadGameButton, newGameButton, saveButton } from '../events'
 import { setTextQueue } from '../textUtils'
 
 let deathCounter = 64
@@ -23,11 +23,15 @@ export const fadeIntoGameOver = () => {
     constants.WIDTH,
     constants.HEIGHT
   )
+  saveButton.active = false
+
   if (deathCounter === 0) {
     CTDLGAME.gameOver = true
-    db.destroy()
+
+    if (CTDLGAME.options.difficulty === 'rogue') db.destroy()
 
     newGameButton.active = true
+    loadGameButton.active = true
 
     stopMusic()
     changeVolume(1)
